Extract `calculateControllerRoots` into project util
