+++ src/util/index.ts
@@ -2,5 +2,6 @@ export * from "./ast"
 export * from "./decorators"
 export * from "./fs"
 export * from "./npm"
+export * from "./project"
 export * from "./properties"
 export * from "./string"
